refactored discovery back into the main protocol class.
