+++ TpLinkConsole/Program.cs
@@ -43,7 +43,7 @@ try
     //await client.SetRelayStateAsync(true, cts.Token);
     //var d = await client.SetLocationAsync(234333 / 10000 * -1, 1800000 / 10000, cts.Token);
     //var e = await client.GetMonthStatsAsync(2023, cts.Token);
-    await client.DiscoverDevicesAsync(cts.Token);
+    var r = await client.DiscoverDevicesAsync(cts.Token);
 }
 catch (SmartHomeException ex)
 {
